move components to model/components
