+++ hydromt/model/components/geoms.py
@@ -19,10 +19,10 @@ from geopandas import GeoDataFrame, GeoSeries
 from geopandas.testing import assert_geodataframe_equal
 from shapely.geometry import box
 
-from hydromt.components.base import ModelComponent
-from hydromt.components.spatial import SpatialModelComponent
 from hydromt.hydromt_step import hydromt_step
 from hydromt.metadata_resolver import ConventionResolver
+from hydromt.model.components.base import ModelComponent
+from hydromt.model.components.spatial import SpatialModelComponent
 
 if TYPE_CHECKING:
     from hydromt.model.model import Model
