+++ tests/conftest.py
@@ -26,9 +26,6 @@ from hydromt.plugins import Plugins
 dask_config.set(scheduler="single-threaded")
 
 from hydromt._typing import SourceMetadata
-from hydromt.components.config import ConfigComponent
-from hydromt.components.spatial import SpatialModelComponent
-from hydromt.components.vector import VectorComponent
 from hydromt.data_adapter.geodataframe import GeoDataFrameAdapter
 from hydromt.data_adapter.geodataset import GeoDatasetAdapter
 from hydromt.data_catalog import DataCatalog
@@ -36,6 +33,9 @@ from hydromt.drivers import GeoDataFrameDriver, RasterDatasetDriver
 from hydromt.drivers.geodataset.geodataset_driver import GeoDatasetDriver
 from hydromt.gis import utils
 from hydromt.metadata_resolver import MetaDataResolver
+from hydromt.model.components.config import ConfigComponent
+from hydromt.model.components.spatial import SpatialModelComponent
+from hydromt.model.components.vector import VectorComponent
 from hydromt.root import ModelRoot
 
 dask_config.set(scheduler="single-threaded")
